perf: Tighten and fix the grouping condition
Patchwork-id: 127320
O-Subject: [PATCH RHEL6 BZ1265262 015/210] perf: Tighten (and fix) the grouping condition
Bugzilla: 1265262
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1265262
https://bugzilla.redhat.com/show_bug.cgi?id=1265262

upstream
========
commit c3c87e770458aa004bd7ed3f29945ff436fd6511
Author: Peter Zijlstra <peterz@infradead.org>
Date: Fri Jan 23 11:19:48 2015 +0100

description
===========
The fix from 9fc81d87420d ("perf: Fix events installation during
moving group") was incomplete in that it failed to recognise that
creating a group with events for different CPUs is semantically
broken -- they cannot be co-scheduled.

Furthermore, it leads to real breakage where, when we create an event
for CPU Y and then migrate it to form a group on CPU X, the code gets
confused where the counter is programmed -- triggered in practice
as well by me via the perf fuzzer.

Fix this by tightening the rules for creating groups. Only allow
grouping of counters that can be co-scheduled in the same context.
This means for the same task and/or the same cpu.

Fixes: 9fc81d87420d ("perf: Fix events installation during moving group")
Signed-off-by: Aristeu Rozanski <aris@redhat.com>