perf tools: Do not attempt to run perf-read-vdso32 if it wasn't built
Patchwork-id: 127344
O-Subject: [PATCH RHEL6 BZ1265262 039/210] perf tools: Do not attempt to run perf-read-vdso32 if it wasn't built
Bugzilla: 1265262
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1265262
https://bugzilla.redhat.com/show_bug.cgi?id=1265262

upstream
========
commit 46b1fa85ff5a2e03423770b3931b97266e8ac6cf
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Oct 23 13:45:24 2014 +0300

description
===========
popen() causes an error message to print if perf-read-vdso32 does not
run.  Avoid that by not trying to run it if it was not built.  Ditto
perf-read-vdsox32.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>