perf annotate: Support source line numbers in annotate
Patchwork-id: 127383
O-Subject: [PATCH RHEL6 BZ1265262 078/210] perf annotate: Support source line numbers in annotate
Bugzilla: 1265262
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1265262
https://bugzilla.redhat.com/show_bug.cgi?id=1265262

upstream
========
commit e592488c01d51763de847fcecb3d969231a483a9
Author: Andi Kleen <ak@linux.intel.com>
Date: Wed Nov 12 18:05:26 2014 -0800

description
===========
With srcline key/sort'ing it's useful to have line numbers in the
annotate window. This patch implements this.

Use objdump -l to request the line numbers and save them in the line
structure. Then the browser displays them for source lines.

The line numbers are not displayed by default, but can be toggled on
with 'k'

There is one unfortunate problem with this setup. For lines not
containing source and which are outside functions objdump -l reports
line numbers off by a few: it always reports the first line number in
the next function even for lines that are outside the function.

I haven't found a nice way to detect/correct this. Probably objdump has
to be fixed.

See https://sourceware.org/bugzilla/show_bug.cgi?id=16433

The line numbers are still useful even with these problems, as most are
correct and the ones which are not are nearby.

v2: Fix help text. Handle (discriminator...) output in objdump.
Left align the line numbers.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>