perf evlist: Do not poll events that use the system_wide flag
Patchwork-id: 127387
O-Subject: [PATCH RHEL6 BZ1265262 082/210] perf evlist: Do not poll events that use the system_wide flag
Bugzilla: 1265262
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1265262
https://bugzilla.redhat.com/show_bug.cgi?id=1265262

upstream
========
commit f90d194a867a5a1db51789b8c4e99bcd196be6bb
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue Nov 11 16:16:39 2014 +0200

description
===========
The system_wide flag causes a selected event to be opened always without
a pid.

Consequently it will never get a POLLHUP, but it is used for tracking in
combination with other events, so it should not need to be polled
anyway.

Therefore don't add it for polling.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>