perf callchain: Make get_srcline fall back to sym+offset
Patchwork-id: 127393
O-Subject: [PATCH RHEL6 BZ1265262 089/210] perf callchain: Make get_srcline fall back to sym+offset
Bugzilla: 1265262
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1265262
https://bugzilla.redhat.com/show_bug.cgi?id=1265262

upstream
========
commit 85c116a6cb91a5c09b7a6c95ffc6a6cbd32cd237
Author: Andi Kleen <ak@linux.intel.com>
Date: Wed Nov 12 18:05:27 2014 -0800

description
===========
When the source line is not found fall back to sym + offset.  This is
generally much more useful than a raw address.

For this we need to pass in the symbol from the caller.

For some callers it's awkward to compute, so we stay at the old
behaviour.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>