perf report: Get rid of report__inc_stat
Patchwork-id: 127476
O-Subject: [PATCH RHEL6 BZ1265262 173/210] perf report: Get rid of report__inc_stat()
Bugzilla: 1265262
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1265262
https://bugzilla.redhat.com/show_bug.cgi?id=1265262

upstream
========
commit 590cd344e2099c7b040b29d3a711b4c26358def5
Author: Namhyung Kim <namhyung@kernel.org>
Date: Mon Dec 22 13:44:09 2014 +0900

description
===========
The report__inc_stat() function collects the number of hist entries in
the session in order to calculate the max size of the progess bar.

It'd be better if it does it during the addition of hist entries so that
it can be used by other places too.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>