perf annotate: Fix fallback to unparsed disassembler line
Patchwork-id: 127515
O-Subject: [PATCH RHEL6 BZ1265262 210/210] perf annotate: Fix fallback to unparsed disassembler line
Bugzilla: 1265262
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1265262
https://bugzilla.redhat.com/show_bug.cgi?id=1265262

upstream
========
commit 3995614d9b0320e10ce202836c8477e1bcf1a2d4
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Mar 5 15:27:28 2015 -0300

description
===========
When annotating source/disasm lines the perf tools parse the output of
objdump, trying to provide augmented output that allows navigating
jumps, calls, etc.

But when a line output by objdump can't be parsed the annotation code
falls back to just presenting the unparsed line.

When fixing a leak in the 0fb9f2aab738 commit ("perf annotate: Fix
memory leaks in LOCK handling") we failed to take that into account and
instead tried to free one of the data structures that should be freed
only when successfully allocated, oops, segfault.

There was a change in the way the objdump output for lock prefixed
instructions is formatted that lead the relevant parser to fail to grok
it.

At least RHEL7 works ok, but Fedora 20 segfaults.

Fix it by making the ins__delete() destructor work like the most basic
destructor: free().

Namely make it accept a NULL pointer and when handling it just do
nothing.

Further investigation is needed to figure out the nature of the objdump
output change so as to make the parser grok it.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>