keys: Reinstate EPERM for a key type name beginning with a '.'
Patchwork-id: 127559
O-Subject: [RHEL6 PATCH 23/25] KEYS: Reinstate EPERM for a key type name beginning with a '.'
Bugzilla: 1282745
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>

Reinstate the generation of EPERM for a key type name beginning with a '.' in
a userspace call.  Types whose name begins with a '.' are internal only.

The test was removed by:

	commit a4e3b8d79a5c6d40f4a9703abf7fe3abcc6c3b8d
	Author: Mimi Zohar <zohar@linux.vnet.ibm.com>
	Date:   Thu May 22 14:02:23 2014 -0400
	Subject: KEYS: special dot prefixed keyring name bug fix

I think we want to keep the restriction on type name so that userspace can't
add keys of a special internal type.

Note that removal of the test causes several of the tests in the keyutils
testsuite to fail.

Signed-off-by: David Howells <dhowells@redhat.com>
Acked-by: Vivek Goyal <vgoyal@redhat.com>
cc: Mimi Zohar <zohar@linux.vnet.ibm.com>
Original-commit: 54e2c2c1a9d6cbb270b0999a38545fa9a69bee43
RH-Bugzilla: 1282745
Signed-off-by: Aristeu Rozanski <aris@redhat.com>