Don't leak a key reference if request_key tries to use a revoked keyring
Patchwork-id: 127561
O-Subject: [RHEL6 PATCH 25/25] Don't leak a key reference if request_key() tries to use a revoked keyring
Bugzilla: 1188442 1282745
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Dave Wysochanski <dwysocha@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>

From: David Jeffery <djeffery@redhat.com>

If a request_key() call to allocate and fill out a key attempts to insert the
key structure into a revoked keyring, the key will leak, using memory and part
of the user's key quota until the system reboots. This is from a failure of
construct_alloc_key() to decrement the key's reference count after the attempt
to insert into the requested keyring is rejected.

key_put() needs to be called in the link_prealloc_failed callpath to ensure
the unused key is released.

Signed-off-by: David Jeffery <djeffery@redhat.com>
Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: James Morris <james.l.morris@oracle.com>
Original-commit: d0709f1e66e8066c4ac6a54620ec116aa41937c0
RH-Bugzilla: 1282745, 1188442
Signed-off-by: Aristeu Rozanski <aris@redhat.com>