nct6775: Allocate attributes dynamically from templates
Patchwork-id: 128288
O-Subject: [RHEL6.8 PATCH 22/52] hwmon: (nct6775) Allocate attributes dynamically from templates
Bugzilla: 1260117
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Resolves BZ:
https://bugzilla.redhat.com/show_bug.cgi?id=1260117

Upstream status:
This is a backport of the following from the linux tree...

commit f73cf632dfb43d6236210aa38038cc91fd053ff6
Author: Guenter Roeck <linux@roeck-us.net>
Date:   Mon Mar 18 09:22:50 2013 -0700

    hwmon: (nct6775) Allocate attributes dynamically from templates

    Static attribute allocation is large and very repetitive.
    Allocate attributes and attribute groups dynamically instead.
    This reduces the size of the driver source by more than 600 lines,
    and object size by more than 20k (more than 30%).

    Signed-off-by: Guenter Roeck <linux@roeck-us.net>

Upstream differences:
Upstream's struct attribute_group has a function pointer, is_visible(),
with a return type of umode_t, but RHEL6's is of mode_t. Upstream's
587a1f1659e8b330b873 made this change. Similarly upstream's struct attribute
has a member, mode, of type umode_t, but RHEL6's is of mode_t. Upstream's
9104e427f3e21ddb380d made this change. Backporting these two commits would
require the backport of upstream's 48176a973d65572e61d0 for completeness.
But doing this affects kABI, and is more impactful than it's worth, as
umode_t is an 'unsigned short', and mode_t is an 'unsigned short'
(if 32-bit kernel) or 'unsigned int' (if 64-bit kernel).

Signed-off-by: Aristeu Rozanski <aris@redhat.com>