use dev_get_platdata
Patchwork-id: 128296
O-Subject: [RHEL6.8 PATCH 30/52] hwmon: use dev_get_platdata()
Bugzilla: 1260117
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Resolves BZ:
https://bugzilla.redhat.com/show_bug.cgi?id=1260117

Upstream status:
This is a partial backport of the following from the linux tree...

commit a8b3a3a53f9a814e9938ea9cc179086ff5c0a387
Author: Jingoo Han <jg1.han@samsung.com>
Date:   Tue Jul 30 17:13:06 2013 +0900

    hwmon: use dev_get_platdata()

    Use the wrapper function for retrieving the platform data instead of
    accessing dev->platform_data directly.

    Signed-off-by: Jingoo Han <jg1.han@samsung.com>
    Signed-off-by: Guenter Roeck <linux@roeck-us.net>

Upstream differences:
Only backported the hunk affecting drivers/hwmon/nct6775.c, as a number of
the other affected files are not in RHEL6, and some of those that are in
RHEL6 don't reference dev->platform_data. Didn't seem worth the bother for
what is at heart a cosmetic change.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>