nct6775: Convert to use hwmon_device_register_with_groups
Patchwork-id: 128300
O-Subject: [RHEL6.8 PATCH 34/52] hwmon: (nct6775) Convert to use hwmon_device_register_with_groups
Bugzilla: 1260117
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Resolves BZ:
https://bugzilla.redhat.com/show_bug.cgi?id=1260117

Upstream status:
This is a backport of the following from the linux tree...

commit 615fc8cb0f900c91ee50f0da0a1821a7c6823671
Author: Guenter Roeck <linux@roeck-us.net>
Date:   Sat Jul 6 09:43:30 2013 -0700

    hwmon: (nct6775) Convert to use hwmon_device_register_with_groups

    Signed-off-by: Guenter Roeck <linux@roeck-us.net>

Upstream differences:
The function hwmon_device_register_with_groups() was introduced by upstream's
bab2243ce1897865e31e, which isn't in RHEL6. And trying to backport it results
in the need for a number of other commits and the suffering of kABI issues and
the pain of other things. So I stuck with hwmon_device_register().

Kept show_name() and DEVICE_ATTR(name, ...), since that is still in RHEL7 and
the customer may be expecting to see that..

Replaced nct6775_device_remove_files() by a call to sysfs_remove_groups().

Signed-off-by: Aristeu Rozanski <aris@redhat.com>