hpsa: clean up aborts
Patchwork-id: 128635
O-Subject: [RHEL6.8 e-stor PATCH V2 05/58] hpsa: clean up aborts
Bugzilla: 1247729
RH-Acked-by: Maurizio Lombardi <mlombard@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Stephen Cameron <stephenmcameron@gmail.com>

V2 fixes use of atomic_add_unless()

Cherry-pick of upstream commit with modifications for conflicts:
https://git.kernel.org/cgit/linux/kernel/git/torvalds/linux.git/commit/?id=9b5c48c28f5651f5e36150131bd7b29cc85a3512

In both cases, conflict was solved by taking the changes in the commit that
matched the change being done to upstream code.  This was in
hpsa_init_one() for the atomic_set(), and declaration of *ext_target_mode.

In addition, atomic_add_unless(&val, -1, 0) is used in place off the
atomic_dec_if_positive() call in wait_for_available_abort_cmd(), with
the return value being compared to "!=0" instead of ">=0" to handle case
where "val" is zero.  For such a case, atomic_add_unless() will return zero,
while atomic_dec_if_positive() would return -1, with both leaving "val" set
to zero.   atomic_dec_if_positive() does not exist in RHEL6.

Below is text from the original commit:

Do not send aborts to logical devices that do not support aborts

Instead of relying on what the Smart Array claims for supporting logical
drives, simply try an abort and see how it responds at device discovery
time.  This way devices that do support aborts (e.g. MSA2000) can work
and we do not waste time trying to send aborts to logical drives that do
not support them (important for high IOPS devices.)

While rescanning devices only test whether devices support aborts
the first time we encounter a device rather than every time.

Some Smart Arrays required aborts to be sent with tags in
the wrong endian byte order.  To avoid having to know about
this, we would send two aborts with tags with each endian order.
On high IOPS devices, this turns out to be not such a hot idea.
So we now have a list of the devices that got the tag backwards,
and we only send it one way.

If all available commands are outstanding and the abort handler
is invoked, the abort handler may not be able to allocate a command
and may busy-wait excessivly.  Reserve a small number of commands
for the abort handler and limit the number of concurrent abort
requests to the number of reserved commands.

Reviewed-by: Scott Teel <scott.teel@pmcs.com>
Reviewed-by: Kevin Barnett <kevin.barnett@pmcs.com>
Reviewed-by: Tomas Henzl <thenzl@redhat.com>
Reviewed-by: Hannes Reinecke <hare@Suse.de>
Signed-off-by: Don Brace <don.brace@pmcs.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: James Bottomley <JBottomley@Odin.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>