Remove the extra validity check on nr_pages
Patchwork-id: 129163
O-Subject: [PATCH RHEL6 BZ1291071 084/315] perf: Remove the extra validity check on nr_pages
Bugzilla: 1291071
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1291071
https://bugzilla.redhat.com/show_bug.cgi?id=1291071

upstream
========
commit 74390aa5567827add5058a3b26eff0ed06a629ba
Author: Kaixu Xia <xiakaixu@huawei.com>
Date: Tue Jan 27 17:55:12 2015 +0800

description
===========
The function is_power_of_2() also do the check on nr_pages, so the first
check performed is unnecessary. On the other hand, the key point is to
ensure @nr_pages is a power-of-two number and mostly @nr_pages is a
nonzero value, so in the most cases, the function is_power_of_2() will
be called.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>