tools lib traceevent: Make plugin options either string or boolean
Patchwork-id: 129282
O-Subject: [PATCH RHEL6 BZ1291071 239/315] tools lib traceevent: Make plugin options either string or boolean
Bugzilla: 1291071
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1291071
https://bugzilla.redhat.com/show_bug.cgi?id=1291071

upstream
========
commit 5dbcfd930e430b33fb4dfded9c1af0cf432772b9
Author: Steven Rostedt <srostedt@redhat.com>
Date: Tue Mar 24 09:57:54 2015 -0400

description
===========
When a plugin option is defined, by default it is a boolean (true or false).

If the option is something else, then it needs to set its "value" field to
a default string other than NULL (can be just "").

If the value is not set then the option is considered boolean, and the
updating of the option value will be handled accordingly.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>