tools lib traceevent kbuffer: Remove extra update to data pointer in PADDING
Patchwork-id: 129311
O-Subject: [PATCH RHEL6 BZ1291071 240/315] tools lib traceevent kbuffer: Remove extra update to data pointer in PADDING
Bugzilla: 1291071
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1291071
https://bugzilla.redhat.com/show_bug.cgi?id=1291071

upstream
========
commit c5e691928bf166ac03430e957038b60adba3cf6c
Author: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Date: Tue Mar 24 09:57:55 2015 -0400

description
===========
When a event PADDING is hit (a deleted event that is still in the ring
buffer), translate_data() sets the length of the padding and also updates
the data pointer which is passed back to the caller.

This is unneeded because the caller also updates the data pointer with
the passed back length. translate_data() should not update the pointer,
only set the length.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>