tools lib traceevent: Honor operator priority
Patchwork-id: 129388
O-Subject: [PATCH RHEL6 BZ1291071 275/315] tools lib traceevent: Honor operator priority
Bugzilla: 1291071
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1291071
https://bugzilla.redhat.com/show_bug.cgi?id=1291071

upstream
========
commit 3201f0dc42f7fad9387afc4692cea3d0c730cba2
Author: Namhyung Kim <namhyung@kernel.org>
Date: Mon Apr 6 14:36:16 2015 +0900

description
===========
Currently it ignores operator priority and just sets processed args as a
right operand.  But it could result in priority inversion in case that
the right operand is also a operator arg and its priority is lower.

For example, following print format is from new kmem events.

  "page=%p", REC->pfn != -1UL ? (((struct page *)(0xffffea0000000000UL)) + (REC->pfn)) : ((void *)0)

But this was treated as below:

  REC->pfn != ((null - 1UL) ? ((struct page *)0xffffea0000000000UL + REC->pfn) : (void *) 0)

In this case, the right arg was '?' operator which has lower priority.
But it just sets the whole arg so making the output confusing - page was
always 0 or 1 since that's the result of logical operation.

With this patch, it can handle it properly like following:

  ((REC->pfn != (null - 1UL)) ? ((struct page *)0xffffea0000000000UL + REC->pfn) : (void *) 0)

[ Replaced 'swap' with 'rotate' in a comment as requested by Steve and agreed by Namhyung ]

Signed-off-by: Aristeu Rozanski <aris@redhat.com>