tools: Fix cross-endian analysis
Patchwork-id: 129333
O-Subject: [PATCH RHEL6 BZ1291071 293/315] perf tools: Fix cross-endian analysis
Bugzilla: 1291071
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1291071
https://bugzilla.redhat.com/show_bug.cgi?id=1291071

upstream
========
commit 1060ab857f64f33a9445881fd31fa91470aeb622
Author: David Ahern <david.ahern@oracle.com>
Date: Thu Apr 9 16:15:46 2015 -0400

description
===========
Trying to analyze a big endian data file on little endian system fails
with the error:

  0xa9b40 [0x70]: failed to process type: 9

The problem is that header parsing is not done correctly because the
file attributes are not swapped. Make it so. With this patch able to
analyze a sparc64 data file on x86_64.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>