gfs2: Reintroduce a timeout in function gfs2_gl_hash_clear
Patchwork-id: 129423
O-Subject: [RHEL6.8 fs PATCH] GFS2: Reintroduce a timeout in function gfs2_gl_hash_clear
Bugzilla: 1291458
RH-Acked-by: Andrew Price <anprice@redhat.com>
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

Hi,

This patch is for bugzilla bug #1291458 - GFS2: umount-related glock hangs no longer timeout

The upstream patch is here:
https://git.kernel.org/cgit/linux/kernel/git/gfs2/linux-gfs2.git/commit/fs/gfs2?h=for-next&id=5f63e3c3599d754dae6448a0e83e42a88809737f

A kernel was built here:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=10233741

It's been included in many of my test kernels in the past six months,
and works properly on the many systems I've had it on.

Patch description:
GFS2: Reintroduce a timeout in function gfs2_gl_hash_clear

At some point in the past, we used to have a timeout when GFS2 was
unmounting, trying to clear out its glocks. If the timeout expires,
it would dump the remaining glocks to the kernel messages so that
developers can debug the problem. That timeout was eliminated,
probably by accident. This patch reintroduces it.

rhbz#1291458