gfs2: keep offset when splitting dir leaf blocks
Patchwork-id: 129472
O-Subject: [RHEL6.8 BZ1151962 PATCH 1/2] gfs2: keep offset when splitting dir leaf blocks
Bugzilla: 1151962
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Upstream: 47378d02fb9931f54c5812ec1ac2fb94e5d403a7

Currently, when gfs2 splits a directory leaf block, the dirents that
need to be copied to the new leaf block are packed into the start of it.
This is good for space efficiency. However, if gfs2 were to copy those
dirents into the exact same offset in the new leaf block as they had in
the old block, it would be able to generate a readdir cookie based on
the dirent location, that would be guaranteed to be unique up well past
where the current code is statistically almost guaranteed to have
collisions. So, gfs2 now keeps the dirent's offset in the block the
same when it copies it to the new leaf block.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>