ib/mlx4: Always use the correct port for mirrored multicast attachments
Patchwork-id: 130164
O-Subject: [RHEL-6.8 PATCH 321/483] IB/mlx4: Always use the correct port for mirrored multicast attachments
Bugzilla: 1120022 1142056 1192551 1246059 1253021 1271492
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Moni Shoua <monis@mellanox.com>

When attaching a QP to a multicast address in bonded mode, there was an
assumption that the port of the QP must be #1. This assumption isn't the
case under the flow which enables maximal usage of the physical ports.

Fix it by always checking the port of the original flow and create the
mirrored flow on the other port.

Fixes: c6215745b66a ('IB/mlx4: Load balance ports in port aggregation mode')
Signed-off-by: Moni Shoua <monis@mellanox.com>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 824c25c1abe70a527646056f6911d181facde9cc)

Signed-off-by: Donald Dutile <ddutile@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>