ib/mlx4: Request alias GUID on demand
Patchwork-id: 130117
O-Subject: [RHEL-6.8 PATCH 427/483] IB/mlx4: Request alias GUID on demand
Bugzilla: 1120022 1142056 1192551 1246059 1253021 1271492
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Yishai Hadas <yishaih@mellanox.com>

Request GIDs from the SM on demand, i.e., when a VF actually needs them,
and release them when the GIDs are no longer in use.

In cloud environments, this is useful for GID migrations, in which a
GID is assigned to a VF on the destination HCA, while the VF on the
source HCA is shutdown (but the GID was not administratively released).

Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit ee59fa0d7e9af130bfc1b75524e04c101670bd5e)

Signed-off-by: Donald Dutile <ddutile@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>