x86/intel/lbr: Limit LBR accesses to TOS in callstack mode
Patchwork-id: 130395
O-Subject: [PATCH RHEL6.8 BZ1189317 15/32] perf/x86/intel/lbr: Limit LBR accesses to TOS in callstack mode
Bugzilla: 1189317
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1189317
https://bugzilla.redhat.com/show_bug.cgi?id=1189317

upstream
========
commit 90405aa02247c1a6313c33e2253f9fd2299ae60b
Author: Andi Kleen <ak@linux.intel.com>
Date: Wed May 27 21:13:18 2015 -0700

description
===========
In callstack mode the LBR is not a ring buffer, but a stack that grows up
and down. This means in  this case we don't need to access all LBRs, only the
ones up to TOS. Do this optimization for the normal LBR read, and the context
switch save/restore code. For save/restore it can be done unconditionally, as
it only runs when call stack mode is active.

This recovers some of the cost of going to 32 LBRs on Skylake.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>