annotate: Compute IPC and basic block cycles
Patchwork-id: 130405
O-Subject: [PATCH RHEL6.8 BZ1189317 26/32] perf annotate: Compute IPC and basic block cycles
Bugzilla: 1189317
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1189317
https://bugzilla.redhat.com/show_bug.cgi?id=1189317

upstream
========
commit 30e863bb6f708c0abd422fbb0e6b295f5ee6407b
Author: Andi Kleen <ak@linux.intel.com>
Date: Sat Jul 18 08:24:50 2015 -0700

description
===========
Compute the IPC and the basic block cycles for the annotate display.

IPC is computed by counting the instructions, and then dividing the
accounted cycles by that count.

The actual IPC computation can only be done at annotate time, because we
need to parse the objdump output first to know the number of
instructions in the basic block.

The cycles/IPC are also put into the perf function annotation so that
the display code can show them.

Again basic block overlaps are not handled, with the longest winning,
but there are some heuristics to hide the IPC when the longest is not
the most common.

v2: Compute IPC correctly.
Signed-off-by: Aristeu Rozanski <aris@redhat.com>