vfs: split generic splice code from i_mutex locking
Patchwork-id: 130503
O-Subject: [RHEL6.8 PATCH v2 1/2] vfs: split generic splice code from i_mutex locking
Bugzilla: 697750
RH-Acked-by: Dave Chinner <dchinner@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

This patch is a backport of upstream submission:
http://oss.sgi.com/archives/xfs/2011-08/msg00120.html

XFS holds locks that should be nested inside the inode->i_mutex when
generic_file_splice_write is called. This function takes the
i_mutex, and so we get a lock inversion that triggers lockdep
warnings and has been found to cause real deadlocks.

XFS does not need the splice code to take the i_mutex to do the page
cache manipulation, so modify the generic splice code to use an
actor function for the code that currently requires the i_mutex to
be taken.  Convert generic_file_splice_write() to use this new
interface supplying a generic actor function that performs the same
actions as the existing code.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>