cifs: Make echo interval tunable
Patchwork-id: 131048
O-Subject: [RHEL6 PATCH fs] [bz1234960]cifs: Make echo interval tunable
Bugzilla: 1234960
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Currently the echo interval is set to 60 seconds using a macro. This
setting determines the interval at which echo requests are sent to the
server on an idling connection. This setting also affects the time
required for a connection to an unresponsive server to timeout.

Making this setting a tunable allows users to control the echo interval
times as well as control the time after which the connecting to an
unresponsive server times out.

To set echo interval, pass the echo_interval=n mount option.

Upstream Status:
Included in the for-next branch of the linux-cifs tree.
commit: ba068927b00604fec0210198ce461a4cb132f41f

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>