rpcrdma: Fix to XDR page base interpretation in marshalling logic
Patchwork-id: 131450
O-Subject: [RHEL-6.8 NFSoRDMA v2 003/106] RPCRDMA: Fix to XDR page base interpretation in marshalling logic.
Bugzilla: 1271499
RH-Acked-by: Sachin Prabhu <sprabhu@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>

From: Tom Tucker <tom@ogc.us>

The RPCRDMA marshalling logic assumed that xdr->page_base was an
offset into the first page of xdr->page_list. It is in fact an
offset into the xdr->page_list itself, that is, it selects the
first page in the page_list and the offset into that page.

The symptom depended in part on the rpc_memreg_strategy, if it was
FRMR, or some other one-shot mapping mode, the connection would get
torn down on a base and bounds error. When the badly marshalled RPC
was retransmitted it would reconnect, get the error, and tear down the
connection again in a loop forever. This resulted in a hung-mount. For
the other modes, it would result in silent data corruption. This bug is
most easily reproduced by writing more data than the filesystem
has space for.

This fix corrects the page_base assumption and otherwise simplifies
the iov mapping logic.

Signed-off-by: Tom Tucker <tom@ogc.us>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
(cherry picked from commit bd7ea31b9e8a342be76e0fe8d638343886c2d8c5)
Signed-off-by: Honggang Li <honli@redhat.com>

Signed-off-by: Aristeu Rozanski <aris@redhat.com>