x86: Add 'index' param to get_event_constraint callback
Patchwork-id: 131566
O-Subject: [PATCH RHEL6 BZ1238702 07/17] perf/x86: Add 'index' param to get_event_constraint() callback
Bugzilla: 1238702
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1238702
https://bugzilla.redhat.com/show_bug.cgi?id=1238702

upstream
========
commit 79cba822443a168c8f7f5b853d9c7225a6d5415e
Author: Stephane Eranian <eranian@google.com>
Date: Mon Nov 17 20:06:56 2014 +0100

description
===========
This patch adds an index parameter to the get_event_constraint()
x86_pmu callback. It is expected to represent the index of the
event in the cpuc->event_list[] array. When the callback is used
for fake_cpuc (evnet validation), then the index must be -1. The
motivation for passing the index is to use it to index into another
cpuc array.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>