x86/intel: Limit to half counters when the HT workaround is enabled, to avoid exclusive mode starvation
Patchwork-id: 131571
O-Subject: [PATCH RHEL6 BZ1238702 12/17] perf/x86/intel: Limit to half counters when the HT workaround is enabled, to avoid exclusive mode starvation
Bugzilla: 1238702
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1238702
https://bugzilla.redhat.com/show_bug.cgi?id=1238702

upstream
========
commit c02cdbf60b51b8d98a49185535f5d527a2965142
Author: Stephane Eranian <eranian@google.com>
Date: Mon Nov 17 20:07:02 2014 +0100

description
===========
This patch limits the number of counters available to each CPU when
the HT bug workaround is enabled.

This is necessary to avoid situation of counter starvation. Such can
arise from configuration where one HT thread, HT0, is using all 4 counters
with corrupting events which require exclusion the the sibling HT, HT1.

In such case, HT1 would not be able to schedule any event until HT0
is done. To mitigate this problem, this patch artificially limits
the number of counters to 2.

That way, we can gurantee that at least 2 counters are not in exclusive
mode and therefore allow the sibling thread to schedule events of the
same type (system vs. per-thread). The 2 counters are not determined
in advance. We simply set the limit to two events per HT.

This helps mitigate starvation in case of events with specific counter
constraints such a PREC_DIST.

Note that this does not elimintate the starvation is all cases. But
it is better than not having it.

(Solution suggested by Peter Zjilstra.)

Signed-off-by: Aristeu Rozanski <aris@redhat.com>