gfs2: Clean up reservation removal
Patchwork-id: 131729
O-Subject: [RHEL6.8 fs PATCH v3 1/4] GFS2: Clean up reservation removal
Bugzilla: 1215794
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

From: Steven Whitehouse <swhiteho@redhat.com>

The reservation for an inode should be cleared when it is truncated so
that we can start again at a different offset for future allocations.
We could try and do better than that, by resetting the search based on
where the truncation started from, but this is only a first step.

In addition, there are three callers of gfs2_rs_delete() but only one
of those should really be testing the value of i_writecount. While
we get away with that in the other cases currently, I think it would
be better if we made that test specific to the one case which
requires it.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>

rhbz#1215794

Signed-off-by: Aristeu Rozanski <aris@redhat.com>