gfs2: Extract quota data from reservations structure (revert 5407e24)
Patchwork-id: 131731
O-Subject: [RHEL6.8 fs PATCH v3 2/4] GFS2: Extract quota data from reservations structure (revert 5407e24)
Bugzilla: 1215794
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

This patch basically reverts the majority of patch 5407e24.
That patch eliminated the gfs2_qadata structure in favor of just
using the reservations structure. The problem with doing that is that
it increases the size of the reservations structure. That is not an
issue until it comes time to fold the reservations structure into the
inode in memory so we know it's always there. By separating out the
quota structure again, we aren't punishing the non-quota users by
making all the inodes bigger, requiring more slab space. This patch
creates a new slab area to allocate the quota stuff so it's managed
a little more sanely.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

rhbz#1215794

Signed-off-by: Aristeu Rozanski <aris@redhat.com>