gfs2: Make rgrp reservations part of the gfs2_inode structure
Patchwork-id: 131732
O-Subject: [RHEL6.8 fs PATCH v3 3/4] GFS2: Make rgrp reservations part of the gfs2_inode structure
Bugzilla: 1215794
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

Before this patch, multi-block reservation structures were allocated
from a special slab. This patch folds the structure into the gfs2_inode
structure. The disadvantage is that the gfs2_inode needs more memory,
even when a file is opened read-only. The advantages are: (a) we don't
need the special slab and the extra time it takes to allocate and
deallocate from it. (b) we no longer need to worry that the structure
exists for things like quota management. (c) This also allows us to
remove the calls to get_write_access and put_write_access since we
know the structure will exist.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

rhbz#1215794

Signed-off-by: Aristeu Rozanski <aris@redhat.com>