dlm: Replace nodeid_to_addr with kernel_getpeername
Patchwork-id: 132452
O-Subject: [RHEL6.8 PATCH 2/2] DLM: Replace nodeid_to_addr with kernel_getpeername
Bugzilla: 1264492
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

This patch replaces the call to nodeid_to_addr with a call to
kernel_getpeername. This avoids taking a spinlock because it may
potentially be called from a softirq context.

It has been posted upstream but not yet committed, it's a regression fix and
should reach 6.8

rhbz#1264492

Signed-off-by: Aristeu Rozanski <aris@redhat.com>