ixgbevf: add support for reporting RSS key and hash table for X550
Patchwork-id: 132896
O-Subject: [RHEL6.8 PATCH bz1249245 v4 17/22] ixgbevf: add support for reporting RSS key and hash table for X550
Bugzilla: 1249245
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

This patch extends the reporting of the RSS key and hash table by
adding support for X550 VFs. The difference is that X550 VFs have
their own registers for RSS key and indirection table, so there is
no need to query the PF.

The RSS key and indirection table are stored in the adapter structure
during the configuration of VFRSSRK and VFRETA which in turn can be
used in ethtool for reporting.

The logic for writing VFRETA is also changed to make sure that the
indirection table is reported correctly.

In addition this patch adds defines for the VFRETA entries and number
of VFRSSRK registers as well as some whitespace cleanups.

Reported-by: Vlad Zolotarov <vladz@cloudius-systems.com>
Signed-off-by: Emil Tantilov <emil.s.tantilov@intel.com>
Tested-by: Phil Schmitt <phillip.j.schmitt@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit 9cba434f630a972b47327ae3d014445033166206)

BZ: 1249245
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=10291094
Tested: In beaker. See BZ for details
Signed-off-by: Aristeu Rozanski <aris@redhat.com>