ixgbe: Fix 1G and 10G link stability for X550EM_x SFP+
Patchwork-id: 132951
O-Subject: [RHEL6.8 PATCH bz1249244 v4 077/117] ixgbe: Fix 1G and 10G link stability for X550EM_x SFP+
Bugzilla: 1249244
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Corinna Vinschen <vinschen@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Configures the CS4227 correctly for both 1G and 10G operation,
by moving the code to ixgbe_setup_mac_link_sfp_x550em(). It
needs to be in this function because we need both the module
type and the speed, and this is the only function in the init
flow that knows the speed. In contrast,
ixgbe_setup_sfp_modules_X550em() does not know the speed, so we
can't do anything useful here. This is a fundamental difference
from the previous flow, and is due to the way the CS4227 is
implemented.

Signed-off-by: Mark Rustad <mark.d.rustad@intel.com>
Tested-by: Phil Schmitt <phillip.j.schmitt@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit e23f33367882450c66f7de8805b98ce7665a7ba9)

Bugzilla: 1249244
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=10291094
Tested: In beaker. See BZ for details
Signed-off-by: Aristeu Rozanski <aris@redhat.com>