zfcpdump: Fix collecting of registers
Patchwork-id: 133311
O-Subject: [RHEL6.8 PATCH] [s390] zfcpdump: Fix collecting of registers
Bugzilla: 1303558
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Description
-----------
The CPU registers are not valid in kernel dumps.

With commit a0d6f865 "[s390] add support to start the kernel in
64 bit mode", the zfcpdump prefix register save code in head64.S
now runs in 64 bit mode instead of 32 bit mode.  In 64 bit mode
the "store status" operation stores the registers at a different
address than in 32 bit.  Therefore the saved registers are not
copied correctly into the dump.

Correct the addresses that the "zfcpdump_prefix_array" is filled
with correct prefix register values.

Bugzilla
--------
BZ 1303558
https://bugzilla.redhat.com/show_bug.cgi?id=1303558

Upstream status of the patch
----------------------------
This is a fixup to the commit ID a0d6f865.  The problem does not
apply to the upstream kernel.

Brew
----
https://brewweb.devel.redhat.com/taskinfo?taskID=10425251

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>