virtio_ring: Support DMA APIs
virtio_ring currently sends the device (usually a hypervisor)
physical addresses of its I/O buffers.  This is okay when DMA
addresses and physical addresses are the same thing, but this isn't
always the case.  For example, this never works on Xen guests, and
it is likely to fail if a physical "virtio" device ever ends up
behind an IOMMU or swiotlb.

The immediate use case for me is to enable virtio on Xen guests.
For that to work, we need vring to support DMA address translation
as well as a corresponding change to virtio_pci or to another
driver.

Signed-off-by: Andy Lutomirski <luto@kernel.org>
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>

Conflict patch:
0001-virtio_ring-Support-DMA-APIs.patch

Merge log:
patching file drivers/virtio/Kconfig
patching file drivers/virtio/virtio_ring.c
Hunk #13 merged at 851,855.
patching file tools/virtio/linux/dma-mapping.h
(EScore specific version string)