qla2xxx: Set relogin flag when we fail to queue login requests
Patchwork-id: 134162
O-Subject: [PATCH RHEL 6.8] qla2xxx: Set relogin flag when we fail to queue login requests.
Bugzilla: 1306033
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Chad Dupuis <chad.dupuis@qlogic.com>

scsi commit id: f8b95efcc008b9516ac692381fb1550ada26fdd5

Testing: Patch verified by NetApp.

If we fail to queue an srb for an async login we should set the
relogin flag so it will be retried as the reason for the queuing
failure was most likely transient.  Failure to do this can lead to
failed paths as login is never retried if the relogin flag is not
set.

Signed-off-by: Chad Dupuis <chad.dupuis@qlogic.com>
Signed-off-by: Himanshu Madhani <himanshu.madhani@qlogic.com>
Reviewed-by: Johannes Thumshirn <jthumshirn@suse.de>
Reviewed-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>