pseries: Make 32-bit MSI quirk work on systems lacking firmware support
Patchwork-id: 134628
O-Subject: [RHEL 6.8 PATCH BZ 1303678 v3 2/2] powerpc/pseries: Make 32-bit MSI quirk work on systems lacking firmware support
Bugzilla: 1303678
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Jerome Glisse <jglisse@redhat.com>

backport of upstream commit:

commit f1dd153121dcb872ae6cba8d52bec97519eb7d97
Author: Brian King <brking@linux.vnet.ibm.com>
Date:   Wed May 22 11:07:46 2013 +0000

    Recent commit e61133dda480062d221f09e4fc18f66763f8ecd0 added support
    for a new firmware feature to force an adapter to use 32 bit MSIs.
    However, this firmware is not available for all systems. The hack
    below allows devices needing 32 bit MSIs to work on these systems as
    well.
    It is careful to only enable this on Gen2 slots, which should limit
    this to configurations where this hack is needed and tested to work.

    [Small change to factor out the hack into a separate function -- BenH]

    Signed-off-by: Brian King <brking@linux.vnet.ibm.com>
    Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

I removed the condition of max_bus_speed != PCIE_SPEED_5_0GT because
bus->max_bus_speed is not filled in rhel6 kernel, and there is
too much code in upstream pci infrastructure to backport for just this
field.

Because this hack is relevant to only 1 graphic card on ppc64 arch, I
believe that with respect to the time and effort that is needed to
backport all that code, it is better to just remove this check.

Signed-off-by: Oded Gabbay <oded.gabbay@redhat.com>
Signed-off-by: Aristeu Rozanski <aris@redhat.com>