sb_edac: fix channel/csrow emulation on Broadwell
Patchwork-id: 134846
O-Subject: [RHEL6.8 PATCH] sb_edac: fix channel/csrow emulation on Broadwell
Bugzilla: 1301230
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

BZ#1301230
brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=10504258
upstream: N/A
tested: on a HP Z840 Workstation which can reproduce the crash reliably

This patch fixes a boot time crash in sb_edac in some Broadwell systems.

In RHEL6 we need to keep sysfs compatibility with the old interface which
provides channel/csrow abstraction of the memory present in the system
while upstream moved to a DIMM based model. Because of that, we're stuck
with non-upstream code to keep emulating such interface with new memory
controllers. This patch fixes Broadwell's memory controller support by
properly picking the present TAD devices in the system so the proper
number of channels is accounted for and the respective array is allocated
properly.

Signed-off-by: Aristeu Rozanski <aris@redhat.com>